Moved 2 functions to utils/vcf
