+++ scout/commands/export/variant.py
@@ -4,7 +4,7 @@ import logging
 import os
 import sys
 from pathlib import Path
-from typing import Any, Dict, Iterable, Optional, Tuple
+from typing import Any, Dict, Optional, Tuple
 
 import click
 from flask import current_app
@@ -20,7 +20,7 @@ from scout.export.variant import (
 )
 from scout.server.blueprints.institutes.controllers import variants_to_managed_variants
 from scout.server.extensions import store
-from scout.utils.vcf import build_vcf_header, validate_vcf_line
+from scout.utils.vcf import build_vcf_header, print_vcf, validate_vcf_line
 
 from .export_handler import bson_handler
 from .utils import build_option, category_option, collaborator_option, json_option
@@ -191,39 +191,6 @@ def resolve_case(
     return case_obj["owner"], case_obj
 
 
-def print_vcf(
-    causatives: Iterable[Dict[str, Any]],
-    case_id: Optional[str],
-    build: str,
-    case_obj: Optional[Dict[str, Any]],
-) -> None:
-    """
-    Print causative variants in VCF format.
-
-    If a case_id is provided, the VCF header is extended with FORMAT
-    and per-individual genotype columns.
-    """
-
-    argv = [Path(sys.argv[0]).name] + sys.argv[1:]
-    header = build_vcf_header(
-        build=build, contains_date=True, argv=argv, source=current_app.config["MONGO_DBNAME"]
-    )
-
-    if case_id:
-        header[-1] += "\tFORMAT"
-        for ind in case_obj["individuals"]:
-            header[-1] += "\t" + ind["individual_id"]
-
-    for line in header:
-        click.echo(line)
-
-    for variant_obj in causatives:
-        if variant_string := get_vcf_entry(
-            variant_obj, case_id=case_id, build=build, info_tags={"EXPORT_CATEGORY": "CAUSATIVE"}
-        ):
-            click.echo(variant_string)
-
-
 @click.command("causatives", short_help="Export causative variants")
 @build_option
 @category_option
@@ -297,11 +264,6 @@ def causatives(
             click.echo(line)
         return
 
-    argv = [Path(sys.argv[0]).name] + sys.argv[1:]
-    vcf_header = build_vcf_header(
-        build=build, contains_date=True, argv=argv, source=current_app.config["MONGO_DBNAME"]
-    )
-
     # If case_id is given, print more complete vcf entries, with INFO,
     # and genotypes
     if case_id:
@@ -310,65 +272,4 @@ def causatives(
         for individual in case_obj["individuals"]:
             vcf_header[-1] = vcf_header[-1] + "\t" + individual["individual_id"]
 
-    print_vcf(causatives, case_id, build, case_obj)
-
-
-def get_vcf_entry(
-    variant_obj: dict, case_id: str = None, build: str = "37", info_tags: Optional[dict] = None
-) -> str | None:
-    """
-    Get vcf entry from variant object
-
-    Add any additional INFO tags in a dict info_tags.
-    """
-
-    pos = variant_obj["position"]
-    end = variant_obj.get("end") or pos
-    category = variant_obj["category"]
-    subcat = variant_obj["sub_category"].upper()
-    var_type = "TYPE" if category in ["snv", "cancer"] else "SVTYPE"
-
-    # Build INFO field
-    if category in ["snv", "cancer"] and not variant_obj.get("end"):
-        info_field = f"{var_type}={subcat}"
-    else:
-        info_field = f"END={end};{var_type}={subcat}"
-
-    for key, value in info_tags.items() if info_tags else {}:
-        info_field += f";{key}={value}"
-
-    ref = variant_obj.get("reference") or "N"
-    if ref == ".":
-        ref = "N"
-
-    alt = variant_obj.get("alternative") or "N"
-    if alt in [".", "-", variant_obj["sub_category"]]:
-        alt = f"<{subcat}>" if category == "sv" else "N"
-
-    chrom = variant_obj["chromosome"]
-    if build == "GRCh38":
-        chrom = variant_obj["chromosome"]
-        if chrom == "MT":
-            chrom = "M"
-        chrom = "".join(["chr", chrom])
-
-    filters = ";".join(variant_obj.get("filters", [])) or "."
-    vcf_fields = [
-        chrom,
-        str(pos),
-        variant_obj.get("dbsnp_id", ".") or ".",
-        ref,
-        alt,
-        str(variant_obj.get("quality", ".") or "."),
-        filters,
-        info_field,
-    ]
-
-    if case_id and variant_obj.get("samples"):
-        vcf_fields.append("GT")
-        vcf_fields.extend(sample["genotype_call"] for sample in variant_obj["samples"])
-
-    variant_string = "\t".join(vcf_fields)
-
-    if validate_vcf_line(var_type=var_type, line=variant_string)[0]:
-        return variant_string
+    print_vcf(variants=causatives, build=build, export_category="CAUSATIVE", case_obj=case_obj)
